backend/chatbot/chatbot/requirements.txt
numpy==2.2.4
pandas==2.2.2
matplotlib>=3.5.0  # Use newer version compatible with Python 3.10
scikit-learn==1.6.1
scipy==1.13.1
tensorflow==2.19.0
torch==2.6.0
transformers==4.44.2
nltk
spacy==3.8.5
gensim
langchain==0.3.18
langchain-openai==0.3.5
flask==2.3.3
fastapi==0.110.3
streamlit
opencv-python==4.11.0.86
pillow
scikit-image
seaborn
plotly
bokeh
altair
jupyter
ipython
requests
beautifulsoup4
sqlalchemy==2.0.21
openai==1.61.1
google-generativeai==0.8.4
huggingface-hub==0.28.1
pydantic==2.10.6
python-dotenv==1.0.1
tqdm
llvmlite
numba
pygments
aiohttp
langchain-community==0.3.17
langchain-core==0.3.45
chainlit==1.1.402
google-auth==2.38.0
gradio_client==1.8.0
flask-cors==5.0.1
datasets==3.2.0
peft==0.14.0
accelerate==1.3.0
safetensors==0.4.4
tokenizers==0.19.1
deepface==0.0.93
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
waitress==3.0.2
pdf2image==1.17.0
pytesseract==0.3.13
bitsandbytes==0.45.2
tiktoken==0.8.0